Move take screenshot logic to its own class
